hybrid-knowledge-system/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hybrid-knowledge-system"
version = "0.1.0"
description = "Hybrid Neo4j + Weaviate knowledge system for business intelligence"
readme = "README.md"
authors = [
    {name = "Living Twin", email = "team@livingtwin.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    # Core Dependencies
    "neo4j==5.18.0",
    "weaviate-client==4.5.1", 
    "redis==5.0.1",
    "python-dotenv==1.0.1",
    
    # ML & Embeddings
    "sentence-transformers==2.4.0",
    "openai==1.12.0",
    "tiktoken==0.6.0",
    "transformers==4.38.2",
    "torch>=2.0.0",
    "numpy==1.26.4",
    
    # Data Processing
    "pandas==2.2.1",
    "langchain==0.1.11",
    "langchain-community==0.0.25",
    "langchain-openai==0.0.8",
    "pypdf==4.1.0",
    "python-docx==1.1.0",
    "beautifulsoup4==4.12.3",
    
    # Web & API
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "httpx==0.27.0",
    "aioredis==2.0.1",
    
    # Monitoring & Logging
    "prometheus-client==0.20.0",
    "structlog==24.1.0",
    "rich==13.7.1",
    
    # Utilities
    "pydantic==2.6.3",
    "typer==0.9.0",
    "tqdm==4.66.2",
    "click==8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest==8.1.1",
    "pytest-asyncio==0.23.5", 
    "pytest-cov==4.1.0",
    "black==24.2.0",
    "isort==5.13.2",
    "mypy==1.9.0",
]

[project.urls]
Homepage = "https://github.com/livingtwin/hybrid-knowledge-system"
Repository = "https://github.com/livingtwin/hybrid-knowledge-system"
Issues = "https://github.com/livingtwin/hybrid-knowledge-system/issues"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true